Fix for production builds where we'd still very occasionally see "esri undefined" errors.
Turns out that last fix didn't work all the time. Could still get esri references before Dojo had finished its business. So, now we explicitly remove the temptation to reference things in the View Config and have to write code in the EsriMap component subclass's initializer to set the initial extent. Shame really, but I'm out of ideas. Could refactor better on the Esri Map Component to provide a convenience method to do this, but for this sample for now, it'll do.